+++ breeze-dsl/breeze-dsl-critic-markup/src/main/kotlin/CriticMarkupDslExtensions.kt
@@ -1,7 +1,7 @@
 // Copyright (c) 2019-2020 DragonKnightOfBreeze Windea
 // Breeze is blowing...
 
-@file:Suppress("unused")
+@file:JvmName("CriticMarkupDslExtensions")
 
 package com.windea.breezeframework.dsl.criticmarkup
 
@@ -11,10 +11,9 @@ import com.windea.breezeframework.dsl.criticmarkup.CriticMarkupDsl.*
  * 开始构建[CriticMarkupDsl]。
  */
 @CriticMarkupDslMarker
-inline fun criticMarkupDsl(block: Document.() -> CharSequence): Document {
-	return Document().apply { text = block() }
+inline fun criticMarkupDsl(block: DslDocument.() -> CharSequence): DslDocument {
+	return DslDocument().apply { text = block() }
 }
-
 
 /**
  * 创建一个[CriticMarkupDsl.Addition]。