+++ breeze-dsl/src/main/kotlin/api/Generatable.kt
@@ -4,7 +4,8 @@
 package com.windea.breezeframework.dsl.api
 
 /**
- * 内容可生成的DSL元素。
+ * 可生成内容。
+ * @property generateContent 是否生成内容。
  */
 @DslApiMarker
 interface Generatable {