+++ breeze-dsl/src/main/kotlin/api/Wrappable.kt
@@ -4,7 +4,8 @@
 package com.windea.breezeframework.dsl.api
 
 /**
- * 内容可换行的DSL元素。
+ * 可换行内容。
+ * @property wrapContent 是否换行内容。
  */
 @DslApiMarker
 interface Wrappable {